backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg[binary]>=3.1,<4.0
django-cors-headers>=4.3,<5.0
